Moved formly types to formly folder
