+++ libs/documentation/src/lib/components/formly/multicheckbox/demos/basic/multicheckbox-basic.module.ts
