+++ libs/documentation/src/lib/components/formly/multicheckbox/multicheckbox.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MultiCheckboxBasic } from './demos/basic/multicheckbox-basic.component';
-import { DocumentationExamplesPage } from '../shared/examples-page/examples.component';
-import { DocumentationAPIPage } from '../shared/api-page/docs-api.component';
-import { DocumentationSourcePage } from '../shared/source-page/source.component';
-import { DocumentationTemplatePage } from '../shared/template-page/template.component';
-import { DocumentationComponentsSharedModule, DocumentationDemoList } from '../shared/index';
-import { ComponentWrapperComponent } from '../../shared/component-wrapper/component-wrapper.component';
+import { DocumentationExamplesPage } from '../../shared/examples-page/examples.component';
+import { DocumentationAPIPage } from '../../shared/api-page/docs-api.component';
+import { DocumentationSourcePage } from '../../shared/source-page/source.component';
+import { DocumentationTemplatePage } from '../../shared/template-page/template.component';
+import { DocumentationComponentsSharedModule, DocumentationDemoList } from '../../shared/index';
+import { ComponentWrapperComponent } from '../../../shared/component-wrapper/component-wrapper.component';
 import { MultiCheckboxBasicModule } from './demos/basic/multicheckbox-basic.module';
 
 declare var require: any;
